Implement cartridge as abstract class to facilitate for additional mappers
